// src/dto/prediction/hsg-subject.enum.ts
// Define the enum for the National Excellent Student Award subjects
export enum HsgSubject {
    ANH = "Anh",
    DIA = "Địa",
    HOA = "Hoá",
    LY = "Lý",
    SINH = "Sinh",
    SU = "Sử",
    TIENG_NGA = "Tiếng Nga",
    TIENG_NHAT = "Tiếng Nhật",
    TIENG_PHAP = "Tiếng Pháp",
    TIENG_TRUNG = "Tiếng Trung",
    TIN = "Tin",
    TOAN = "Toán",
    VAN = "Văn",
}