// src/dto/prediction/l3-national-subject.enum.ts
export enum L3NationalSubject {
    CONG_NGHE_CONG_NGHIEP = "Công nghệ công nghiệp",
    CONG_NGHE_NONG_NGHIEP = "Công nghệ nông nghiệp",
    DIA_LY = "Địa",
    GDKTPL = "GDKT&PL",
    HOA_HOC = "Hóa",
    LICH_SU = "Sử",
    NGU_VAN = "Văn",
    SINH_HOC = "Sinh",
    TIENG_ANH = "Anh",
    TIENG_DUC = "Tiếng Đức",
    TIENG_HAN = "Tiếng Hàn",
    TIENG_NGA = "Tiếng Nga",
    TIENG_NHAT = "Tiếng Nhật",
    TIENG_PHAP = "Tiếng Pháp",
    TIENG_TRUNG = "Tiếng Trung",
    TIN_HOC = "Tin",
    TOAN = "Toán",
    VAT_LY = "Lý",
}